Add support for requests package
If the requests package is available, chain will import and use it
and users should expect the requests' response object returned.
I'll have to standardize the response object so that one does not
have to guess what is returned. Either that or make the usage of
the requests package explicit.